# protocol/tx: add cross-entry pointers in parallel with their hashes