--- a/protocol/tx/header.go
+++ b/protocol/tx/header.go
@@ -10,6 +10,10 @@ type header struct {
 		MinTimeMS, MaxTimeMS uint64
 		ExtHash              bc.Hash
 	}
+
+	// Results contains (pointers to) the manifested entries for the
+	// items in body.Results.
+	Results []entry
 }
 
 func (header) Type() string         { return "txheader" }
@@ -17,12 +21,20 @@ func (h *header) Body() interface{} { return h.body }
 
 func (header) Ordinal() int { return -1 }
 
-func newHeader(version uint64, results []bc.Hash, data bc.Hash, minTimeMS, maxTimeMS uint64) *header {
+func newHeader(version uint64, results []entry, data bc.Hash, minTimeMS, maxTimeMS uint64) *header {
 	h := new(header)
 	h.body.Version = version
-	h.body.Results = results
 	h.body.Data = data
 	h.body.MinTimeMS = minTimeMS
 	h.body.MaxTimeMS = maxTimeMS
+
+	if len(results) > 0 {
+		h.Results = results
+		h.body.Results = make([]bc.Hash, 0, len(results))
+		for _, r := range results {
+			h.body.Results = append(h.body.Results, entryID(r))
+		}
+	}
+
 	return h
 }