--- a/protocol/tx/issuance.go
+++ b/protocol/tx/issuance.go
@@ -10,6 +10,10 @@ type issuance struct {
 		ExtHash bc.Hash
 	}
 	ordinal int
+
+	// Anchor is a pointer to the manifested entry corresponding to
+	// body.Anchor.
+	Anchor entry
 }
 
 func (issuance) Type() string           { return "issuance1" }
@@ -17,9 +21,12 @@ func (iss *issuance) Body() interface{} { return iss.body }
 
 func (iss issuance) Ordinal() int { return iss.ordinal }
 
-func newIssuance(anchor bc.Hash, value bc.AssetAmount, data bc.Hash, ordinal int) *issuance {
+func newIssuance(anchor entry, value bc.AssetAmount, data bc.Hash, ordinal int) *issuance {
 	iss := new(issuance)
-	iss.body.Anchor = anchor
+	if anchor != nil {
+		iss.body.Anchor = entryID(anchor)
+		iss.Anchor = anchor
+	}
 	iss.body.Value = value
 	iss.body.Data = data
 	iss.ordinal = ordinal