--- a/protocol/tx/issuance.go
+++ b/protocol/tx/issuance.go
@@ -23,10 +23,8 @@ func (iss issuance) Ordinal() int { return iss.ordinal }
 
 func newIssuance(anchor entry, value bc.AssetAmount, data bc.Hash, ordinal int) *issuance {
 	iss := new(issuance)
-	if anchor != nil {
-		iss.body.Anchor = entryID(anchor)
-		iss.Anchor = anchor
-	}
+	iss.body.Anchor = entryID(anchor)
+	iss.Anchor = anchor
 	iss.body.Value = value
 	iss.body.Data = data
 	iss.ordinal = ordinal