--- a/protocol/tx/map.go
+++ b/protocol/tx/map.go
@@ -13,7 +13,7 @@ import (
 func mapTx(tx *bc.TxData) (headerID bc.Hash, hdr *header, entryMap map[bc.Hash]entry, err error) {
 	entryMap = make(map[bc.Hash]entry)
 
-	addEntry := func(e entry) (id bc.Hash, entry entry, err error) {
+	addEntry := func(e entry) (id bc.Hash, err error) {
 		defer func() {
 			if pErr, ok := recover().(error); ok {
 				err = pErr
@@ -21,35 +21,32 @@ func mapTx(tx *bc.TxData) (headerID bc.Hash, hdr *header, entryMap map[bc.Hash]e
 		}()
 		id = entryID(e)
 		entryMap[id] = e
-		return id, e, err
+		return id, err
 	}
 
 	// Loop twice over tx.Inputs, once for spends and once for
 	// issuances.  Do spends first so the entry ID of the first spend is
 	// available in case an issuance needs it for its anchor.
 
-	var firstSpendID *bc.Hash
+	var firstSpend *spend
 	muxSources := make([]valueSource, len(tx.Inputs))
 
 	for i, inp := range tx.Inputs {
 		if oldSp, ok := inp.TypedInput.(*bc.SpendInput); ok {
-			var spentOutputID bc.Hash
-			spentOutputID, err = ComputeOutputID(&oldSp.SpendCommitment)
-			if err != nil {
-				return
-			}
-			var spID bc.Hash
-			spID, _, err = addEntry(newSpend(spentOutputID, hashData(inp.ReferenceData), i))
+			sp := newSpend(hashData(inp.ReferenceData), i)
+			sp.setSpentInfo(oldSp.AssetAmount, program{VMVersion: oldSp.VMVersion, Code: oldSp.ControlProgram})
+			var id bc.Hash
+			id, err = addEntry(sp)
 			if err != nil {
 				err = errors.Wrapf(err, "adding spend entry for input %d", i)
 				return
 			}
 			muxSources[i] = valueSource{
-				Ref:   spID,
+				Ref:   id,
 				Value: oldSp.AssetAmount,
 			}
-			if firstSpendID == nil {
-				firstSpendID = &spID
+			if firstSpend == nil {
+				firstSpend = sp
 			}
 		}
 	}
@@ -60,17 +57,17 @@ func mapTx(tx *bc.TxData) (headerID bc.Hash, hdr *header, entryMap map[bc.Hash]e
 			// programs are omitted here because they do not contribute to
 			// the body hash of an issuance.
 
-			var nonceHash bc.Hash
+			var nonce entry
 
 			if len(oldIss.Nonce) == 0 {
-				if firstSpendID == nil {
+				if firstSpend == nil {
 					err = fmt.Errorf("nonce-less issuance in transaction with no spends")
 					return
 				}
-				nonceHash = *firstSpendID
+				nonce = firstSpend
 			} else {
-				var trID bc.Hash
-				trID, _, err = addEntry(newTimeRange(tx.MinTime, tx.MaxTime))
+				tr := newTimeRange(tx.MinTime, tx.MaxTime)
+				_, err = addEntry(tr)
 				if err != nil {
 					err = errors.Wrapf(err, "adding timerange entry for input %d", i)
 					return
@@ -80,7 +77,8 @@ func mapTx(tx *bc.TxData) (headerID bc.Hash, hdr *header, entryMap map[bc.Hash]e
 				b := vmutil.NewBuilder()
 				b = b.AddData(oldIss.Nonce).AddOp(vm.OP_DROP).AddOp(vm.OP_ASSET).AddData(assetID[:]).AddOp(vm.OP_EQUAL)
 
-				nonceHash, _, err = addEntry(newNonce(program{1, b.Program}, trID))
+				nonce = newNonce(program{1, b.Program}, tr)
+				_, err = addEntry(nonce)
 				if err != nil {
 					err = errors.Wrapf(err, "adding nonce entry for input %d", i)
 					return
@@ -89,8 +87,9 @@ func mapTx(tx *bc.TxData) (headerID bc.Hash, hdr *header, entryMap map[bc.Hash]e
 
 			val := inp.AssetAmount()
 
+			iss := newIssuance(nonce, val, hashData(inp.ReferenceData), i)
 			var issID bc.Hash
-			issID, _, err = addEntry(newIssuance(nonceHash, val, hashData(inp.ReferenceData), i))
+			issID, err = addEntry(iss)
 			if err != nil {
 				err = errors.Wrapf(err, "adding issuance entry for input %d", i)
 				return
@@ -103,50 +102,52 @@ func mapTx(tx *bc.TxData) (headerID bc.Hash, hdr *header, entryMap map[bc.Hash]e
 		}
 	}
 
-	muxID, _, err := addEntry(newMux(muxSources, program{VMVersion: 1, Code: []byte{byte(vm.OP_TRUE)}}))
+	mux := newMux(program{VMVersion: 1, Code: []byte{byte(vm.OP_TRUE)}})
+	for _, src := range muxSources {
+		// xxx addSource will recompute the hash of entryMap[src.Ref], which is available as src.Ref - fix this
+		mux.addSource(entryMap[src.Ref], src.Value, src.Position)
+	}
+	_, err = addEntry(mux)
 	if err != nil {
 		err = errors.Wrap(err, "adding mux entry")
 		return
 	}
 
-	var results []bc.Hash
+	var results []entry
 
 	for i, out := range tx.Outputs {
-		s := valueSource{
-			Ref:      muxID,
-			Position: uint64(i),
-			Value:    out.AssetAmount,
-		}
-
-		var resultID bc.Hash
 		if vmutil.IsUnspendable(out.ControlProgram) {
 			// retirement
-			resultID, _, err = addEntry(newRetirement(s, hashData(out.ReferenceData), i))
+			r := newRetirement(hashData(out.ReferenceData), i)
+			r.setSource(mux, out.AssetAmount, uint64(i))
+			_, err = addEntry(r)
 			if err != nil {
 				err = errors.Wrapf(err, "adding retirement entry for output %d", i)
 				return
 			}
+			results = append(results, r)
 		} else {
 			// non-retirement
 			prog := program{out.VMVersion, out.ControlProgram}
-			resultID, _, err = addEntry(newOutput(s, prog, hashData(out.ReferenceData), i))
+			o := newOutput(prog, hashData(out.ReferenceData), i)
+			o.setSource(mux, out.AssetAmount, uint64(i))
+			_, err = addEntry(o)
 			if err != nil {
 				err = errors.Wrapf(err, "adding output entry for output %d", i)
 				return
 			}
+			results = append(results, o)
 		}
-
-		results = append(results, resultID)
 	}
 
-	var h entry
-	headerID, h, err = addEntry(newHeader(tx.Version, results, hashData(tx.ReferenceData), tx.MinTime, tx.MaxTime))
+	h := newHeader(tx.Version, results, hashData(tx.ReferenceData), tx.MinTime, tx.MaxTime)
+	headerID, err = addEntry(h)
 	if err != nil {
 		err = errors.Wrap(err, "adding header entry")
 		return
 	}
 
-	return headerID, h.(*header), entryMap, nil
+	return headerID, h, entryMap, nil
 }
 
 func mapBlockHeader(old *bc.BlockHeader) (bhID bc.Hash, bh *blockHeader) {