--- a/protocol/tx/mux.go
+++ b/protocol/tx/mux.go
@@ -8,6 +8,10 @@ type mux struct {
 		Program program
 		ExtHash bc.Hash
 	}
+
+	// Sources contains (pointers to) the manifested entries for each
+	// body.Sources[i].Ref.
+	Sources []entry
 }
 
 func (mux) Type() string         { return "mux1" }
@@ -15,9 +19,18 @@ func (m *mux) Body() interface{} { return m.body }
 
 func (mux) Ordinal() int { return -1 }
 
-func newMux(sources []valueSource, program program) *mux {
+func newMux(program program) *mux {
 	m := new(mux)
-	m.body.Sources = sources
 	m.body.Program = program
 	return m
+}
+
+func (m *mux) addSource(e entry, value bc.AssetAmount, position uint64) {
+	src := valueSource{
+		Ref:      entryID(e),
+		Value:    value,
+		Position: position,
+	}
+	m.body.Sources = append(m.body.Sources, src)
+	m.Sources = append(m.Sources, e)
 }