--- a/protocol/tx/mux.go
+++ b/protocol/tx/mux.go
@@ -26,11 +26,16 @@ func newMux(program program) *mux {
 }
 
 func (m *mux) addSource(e entry, value bc.AssetAmount, position uint64) {
+	m.addSourceID(entryID(e), value, position)
+	m.Sources[len(m.Sources)-1] = e
+}
+
+func (m *mux) addSourceID(sourceID bc.Hash, value bc.AssetAmount, position uint64) {
 	src := valueSource{
-		Ref:      entryID(e),
+		Ref:      sourceID,
 		Value:    value,
 		Position: position,
 	}
 	m.body.Sources = append(m.body.Sources, src)
-	m.Sources = append(m.Sources, e)
+	m.Sources = append(m.Sources, nil)
 }