--- a/protocol/tx/nonce.go
+++ b/protocol/tx/nonce.go
@@ -8,6 +8,10 @@ type nonce struct {
 		TimeRange bc.Hash
 		ExtHash   bc.Hash
 	}
+
+	// TimeRange contains (a pointer to) the manifested entry
+	// corresponding to body.TimeRange.
+	TimeRange *timeRange
 }
 
 func (nonce) Type() string         { return "nonce1" }
@@ -15,9 +19,12 @@ func (n *nonce) Body() interface{} { return n.body }
 
 func (nonce) Ordinal() int { return -1 }
 
-func newNonce(p program, tr bc.Hash) *nonce {
+func newNonce(p program, tr *timeRange) *nonce {
 	n := new(nonce)
 	n.body.Program = p
-	n.body.TimeRange = tr
+	if tr != nil {
+		n.body.TimeRange = entryID(tr)
+		n.TimeRange = tr
+	}
 	return n
 }