--- a/protocol/tx/output.go
+++ b/protocol/tx/output.go
@@ -10,6 +10,10 @@ type output struct {
 		ExtHash        bc.Hash
 	}
 	ordinal int
+
+	// Source contains (a pointer to) the manifested entry corresponding
+	// to body.Source.
+	Source entry
 }
 
 func (output) Type() string         { return "output1" }
@@ -17,11 +21,19 @@ func (o *output) Body() interface{} { return o.body }
 
 func (o output) Ordinal() int { return o.ordinal }
 
-func newOutput(source valueSource, controlProgram program, data bc.Hash, ordinal int) *output {
+func newOutput(controlProgram program, data bc.Hash, ordinal int) *output {
 	out := new(output)
-	out.body.Source = source
 	out.body.ControlProgram = controlProgram
 	out.body.Data = data
 	out.ordinal = ordinal
 	return out
+}
+
+func (o *output) setSource(e entry, value bc.AssetAmount, position uint64) {
+	o.body.Source = valueSource{
+		Ref:      entryID(e),
+		Value:    value,
+		Position: position,
+	}
+	o.Source = e
 }