--- a/protocol/tx/output.go
+++ b/protocol/tx/output.go
@@ -29,11 +29,19 @@ func newOutput(controlProgram program, data bc.Hash, ordinal int) *output {
 	return out
 }
 
+// setSource is for when you have a complete source entry (e.g. a
+// *mux) for an output. When you don't (you only have the ID of the
+// source), use setSourceID, below.
 func (o *output) setSource(e entry, value bc.AssetAmount, position uint64) {
+	o.setSourceID(entryID(e), value, position)
+	o.Source = e
+}
+
+func (o *output) setSourceID(sourceID bc.Hash, value bc.AssetAmount, position uint64) {
 	o.body.Source = valueSource{
-		Ref:      entryID(e),
+		Ref:      sourceID,
 		Value:    value,
 		Position: position,
 	}
-	o.Source = e
+	o.Source = nil
 }