--- a/protocol/tx/retirement.go
+++ b/protocol/tx/retirement.go
@@ -9,6 +9,10 @@ type retirement struct {
 		ExtHash bc.Hash
 	}
 	ordinal int
+
+	// Source contains (a pointer to) the manifested entry corresponding
+	// to body.Source.
+	Source entry
 }
 
 func (retirement) Type() string         { return "retirement1" }
@@ -16,10 +20,18 @@ func (r *retirement) Body() interface{} { return r.body }
 
 func (r retirement) Ordinal() int { return r.ordinal }
 
-func newRetirement(source valueSource, data bc.Hash, ordinal int) *retirement {
+func newRetirement(data bc.Hash, ordinal int) *retirement {
 	r := new(retirement)
-	r.body.Source = source
 	r.body.Data = data
 	r.ordinal = ordinal
 	return r
+}
+
+func (r *retirement) setSource(e entry, value bc.AssetAmount, position uint64) {
+	r.body.Source = valueSource{
+		Ref:      entryID(e),
+		Value:    value,
+		Position: position,
+	}
+	r.Source = e
 }