--- a/protocol/tx/spend.go
+++ b/protocol/tx/spend.go
@@ -9,6 +9,20 @@ type spend struct {
 		ExtHash     bc.Hash
 	}
 	ordinal int
+
+	// SpentOutput contains (a pointer to) the manifested entry
+	// corresponding to body.SpentOutput. It may be nil, in which case
+	// spentInfo must be non-nil.
+	SpentOutput *output
+
+	// spentInfo contains validation-essential data in case the spent
+	// output object is not present (in SpentOutput).
+	spentInfo *spentInfo
+}
+
+type spentInfo struct {
+	bc.AssetAmount
+	program
 }
 
 func (spend) Type() string         { return "spend1" }
@@ -16,10 +30,21 @@ func (s *spend) Body() interface{} { return s.body }
 
 func (s spend) Ordinal() int { return s.ordinal }
 
-func newSpend(spentOutput bc.Hash, data bc.Hash, ordinal int) *spend {
+func newSpend(data bc.Hash, ordinal int) *spend {
 	s := new(spend)
-	s.body.SpentOutput = spentOutput
 	s.body.Data = data
 	s.ordinal = ordinal
 	return s
+}
+
+func (s *spend) setSpentOutput(o *output) {
+	s.body.SpentOutput = entryID(o)
+	s.SpentOutput = o
+}
+
+func (s *spend) setSpentInfo(value bc.AssetAmount, program program) {
+	s.spentInfo = &spentInfo{
+		AssetAmount: value,
+		program:     program,
+	}
 }