The pointers are outside each entry's `body` and `witness`. This is an alternative to the approach in https://github.com/chain/chain/pull/623.