@bobg 2017-03-01
Unfortunately, transaction hashing has slowed down (almost certainly because of redundant hashing of component entries):

```
benchmark                     old ns/op     new ns/op     delta
BenchmarkHashNonemptyTx-8     25921         46452         +79.21%
```

(The `EntryRef` change in https://github.com/chain/chain/pull/623 also slows things down, but by a lesser amount: 23.65%.)

I think this can be addressed in subsequent commits, but I don't plan to check in this one until I've proven out some optimization ideas. Stay tuned.