@kr 2017-03-01
I would argue the extra hashing shouldn't block landing this PR until we've proven it makes a material difference to our overall performance.