@bobg 2017-03-03
rebased